move out of internals & expose createOutputMiddleware
